+++ src/main/java/biz/nellemann/svci/Utils.java
@@ -10,13 +10,16 @@ import java.util.Locale;
 public class Utils {
 
     static public Instant parseDateTime(String stringDate) {
+        if(stringDate == null) {
+            System.err.println("parseDateTime() - null input");
+            return Instant.now();
+        }
         String pattern = "yyyy-MM-dd HH:mm:ss";
         DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern(pattern, Locale.ENGLISH);
         LocalDateTime localDateTime = LocalDateTime.parse(stringDate, dateTimeFormatter);
         ZoneId zoneId = ZoneId.of("UTC");
         ZonedDateTime zonedDateTime = localDateTime.atZone(zoneId);
-        Instant instant = zonedDateTime.toInstant();
-        return instant;
+        return zonedDateTime.toInstant();
     }
 
 }